pyproject.toml
[tool.poetry]
name = "ki-adminpanel"
version = "0.1.0"
description = ""
authors = ["Horst Amper <HorstA@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
customtkinter = "^5.2.2"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
requests = "^2.32.3"
watchdog = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
